requirements.txt
#solo las librerias de google-genai
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
google-auth==2.38.0
google-genai==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
# la libreria de streamlit 
streamlit==1.43.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tzdata==2025.1
#entorno .env
python-dotenv-1.0.1
# postgresql librerias
psycopg2-binary==2.9.10
